refactor : move study group services into a single StudyGroupMemberService.
