+++ src/main/java/com/samsamhajo/deepground/global/config/MongoConfig.java
@@ -9,9 +9,9 @@ import org.springframework.data.mongodb.config.EnableMongoAuditing;
 @Configuration
 @EnableMongoAuditing
 public class MongoConfig {
-
-    @Bean
-    public MongoTransactionManager transactionManager(MongoDatabaseFactory factory) {
-        return new MongoTransactionManager(factory);
-    }
+//
+//    @Bean
+//    public MongoTransactionManager transactionManager(MongoDatabaseFactory factory) {
+//        return new MongoTransactionManager(factory);
+//    }
 }
