--- src/main/java/com/samsamhajo/deepground/studyGroup/service/StudyGroupKickService.java
@@ -1,58 +1,0 @@
-package com.samsamhajo.deepground.studyGroup.service;
-
-import com.samsamhajo.deepground.chat.service.ChatRoomMemberService;
-import com.samsamhajo.deepground.member.entity.Member;
-import com.samsamhajo.deepground.notification.entity.data.StudyGroupNotificationData;
-import com.samsamhajo.deepground.notification.event.NotificationEvent;
-import com.samsamhajo.deepground.studyGroup.dto.StudyGroupKickRequest;
-import com.samsamhajo.deepground.studyGroup.entity.StudyGroup;
-import com.samsamhajo.deepground.studyGroup.entity.StudyGroupMember;
-import com.samsamhajo.deepground.studyGroup.exception.StudyGroupNotFoundException;
-import com.samsamhajo.deepground.studyGroup.repository.StudyGroupMemberRepository;
-import com.samsamhajo.deepground.studyGroup.repository.StudyGroupRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-public class StudyGroupKickService {
-
-  private final StudyGroupRepository studyGroupRepository;
-  private final StudyGroupMemberRepository studyGroupMemberRepository;
-  private final ChatRoomMemberService chatRoomMemberService;
-  private final ApplicationEventPublisher eventPublisher;
-
-  @Transactional
-  public void kickMember(StudyGroupKickRequest request, Member requester) {
-    StudyGroup group = studyGroupRepository.findById(request.getStudyGroupId())
-        .orElseThrow(() -> new StudyGroupNotFoundException(request.getStudyGroupId()));
-
-    if (!group.getCreator().getId().equals(requester.getId())) {
-      throw new IllegalArgumentException("스터디장만 강퇴할 수 있습니다.");
-    }
-
-    if (request.getTargetMemberId().equals(requester.getId())) {
-      throw new IllegalArgumentException("자기 자신은 강퇴할 수 없습니다.");
-    }
-
-    StudyGroupMember target = studyGroupMemberRepository.findByStudyGroupIdAndMemberId(
-            request.getStudyGroupId(), request.getTargetMemberId())
-        .orElseThrow(() -> new IllegalArgumentException("대상 멤버가 스터디에 존재하지 않습니다."));
-
-    // 채팅방 멤버 삭제
-    if (target.getIsAllowed()) {
-      chatRoomMemberService.leaveChatRoom(group.getChatRoom().getId(), target.getMember().getId());
-
-      // 스터디 그룹 강퇴 알림
-      eventPublisher.publishEvent(NotificationEvent.of(
-              request.getTargetMemberId(),
-              StudyGroupNotificationData.kick(group)
-      ));
-    }
-
-    target.softDelete();
-  }
-}
-
