+++ src/main/java/com/samsamhajo/deepground/studyGroup/service/StudyGroupMemberService.java
@@ -1,0 +1,115 @@
+package com.samsamhajo.deepground.studyGroup.service;
+
+import com.samsamhajo.deepground.chat.service.ChatRoomMemberService;
+import com.samsamhajo.deepground.member.entity.Member;
+import com.samsamhajo.deepground.notification.entity.data.StudyGroupNotificationData;
+import com.samsamhajo.deepground.notification.event.NotificationEvent;
+import com.samsamhajo.deepground.studyGroup.dto.StudyGroupKickRequest;
+import com.samsamhajo.deepground.studyGroup.dto.StudyGroupMemberSummary;
+import com.samsamhajo.deepground.studyGroup.entity.StudyGroup;
+import com.samsamhajo.deepground.studyGroup.entity.StudyGroupMember;
+import com.samsamhajo.deepground.studyGroup.exception.StudyGroupNotFoundException;
+import com.samsamhajo.deepground.studyGroup.repository.StudyGroupInviteTokenRepository;
+import com.samsamhajo.deepground.studyGroup.repository.StudyGroupMemberRepository;
+import com.samsamhajo.deepground.studyGroup.repository.StudyGroupRepository;
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class StudyGroupMemberService {
+  private final StudyGroupRepository studyGroupRepository;
+  private final StudyGroupMemberRepository studyGroupMemberRepository;
+  private final ChatRoomMemberService chatRoomMemberService;
+  private final StudyGroupInviteTokenRepository inviteTokenRepository;
+  private final ApplicationEventPublisher eventPublisher;
+
+  @Transactional
+  public List<StudyGroupMemberSummary> getAcceptedMembers(Long studyGroupId) {
+    StudyGroup group = studyGroupRepository.findById(studyGroupId)
+        .orElseThrow(() -> new StudyGroupNotFoundException(studyGroupId));
+
+    return studyGroupMemberRepository.findAcceptedMembersWithInfo(studyGroupId).stream()
+        .map(member -> StudyGroupMemberSummary.builder()
+            .memberId(member.getMember().getId())
+            .nickname(member.getMember().getNickname())
+            .isOwner(group.getCreator().getId().equals(member.getMember().getId()))
+            .joinedAt(member.getCreatedAt())
+            .build())
+        .toList();
+  }
+
+  public List<StudyGroupMemberSummary> getPendingApplicantsAsCreator(Long studyGroupId, Member requester) {
+    StudyGroup studyGroup = studyGroupRepository.findById(studyGroupId)
+        .orElseThrow(() -> new StudyGroupNotFoundException(studyGroupId));
+
+    if (!studyGroup.getCreator().getId().equals(requester.getId())) {
+      throw new IllegalArgumentException("스터디장만 신청자 목록을 조회할 수 있습니다.");
+    }
+
+    List<StudyGroupMember> pending = studyGroupMemberRepository.findAllByStudyGroupIdAndIsAllowedFalse(studyGroupId);
+
+    return pending.stream()
+        .map(m -> StudyGroupMemberSummary.builder()
+            .memberId(m.getMember().getId())
+            .nickname(m.getMember().getNickname())
+            .build())
+        .toList();
+  }
+
+  @Transactional
+  public void requestToJoin(Member member, Long studyGroupId) {
+    StudyGroup studyGroup = studyGroupRepository.findById(studyGroupId)
+        .orElseThrow(() -> new IllegalArgumentException("스터디 그룹이 존재하지 않습니다."));
+
+    // 중복 요청 방지
+    if (studyGroupMemberRepository.existsByMemberAndStudyGroup(member, studyGroup)) {
+      throw new IllegalStateException("이미 참가 요청을 했거나 참여 중입니다.");
+    }
+
+    StudyGroupMember pendingRequest = StudyGroupMember.of(member, studyGroup, false);
+    studyGroupMemberRepository.save(pendingRequest);
+
+    // 스터디 그룹 가입 알림
+    eventPublisher.publishEvent(NotificationEvent.of(
+        studyGroup.getCreator().getId(),
+        StudyGroupNotificationData.join(studyGroup)
+    ));
+
+  }
+
+  @Transactional
+  public void kickMember(StudyGroupKickRequest request, Member requester) {
+    StudyGroup group = studyGroupRepository.findById(request.getStudyGroupId())
+        .orElseThrow(() -> new StudyGroupNotFoundException(request.getStudyGroupId()));
+
+    if (!group.getCreator().getId().equals(requester.getId())) {
+      throw new IllegalArgumentException("스터디장만 강퇴할 수 있습니다.");
+    }
+
+    if (request.getTargetMemberId().equals(requester.getId())) {
+      throw new IllegalArgumentException("자기 자신은 강퇴할 수 없습니다.");
+    }
+
+    StudyGroupMember target = studyGroupMemberRepository.findByStudyGroupIdAndMemberId(
+            request.getStudyGroupId(), request.getTargetMemberId())
+        .orElseThrow(() -> new IllegalArgumentException("대상 멤버가 스터디에 존재하지 않습니다."));
+
+    // 채팅방 멤버 삭제
+    if (target.getIsAllowed()) {
+      chatRoomMemberService.leaveChatRoom(group.getChatRoom().getId(), target.getMember().getId());
+
+      // 스터디 그룹 강퇴 알림
+      eventPublisher.publishEvent(NotificationEvent.of(
+          request.getTargetMemberId(),
+          StudyGroupNotificationData.kick(group)
+      ));
+    }
+
+    target.softDelete();
+  }
+
+}
